Modularize code and add tests
