+++ index.html
@@ -75,6 +75,6 @@
   </div>
 
   <!-- Link to your main JavaScript file -->
-  <script src="game.js"></script>
+  <script type="module" src="game.js"></script>
 </body>
 </html>
